fix: Resolve 15 bugs from third round of application-wide bug hunt
Round 3 findings (down from 130 in round 2):
- Fix invite registration double-accept navigation in LoginPage
- Add inspector notification and guard consultant query in equipment/materials
- Add pagination and status validation to flat materials endpoint
- Fix budget approved_change_orders: SUM→COUNT with matching schema type
- Persist chat action failed status with db.commit before raising
- Fix approval schema field names to match ORM model
- Fix inspection template stage_order field name and add is_active
- Include global templates (project_id IS NULL) in checklist queries
- Wrap blocking Gemini call with asyncio.to_thread in defects
- Set meeting status after sending invitations
- Fix ICS download to use authenticated fetch instead of window.open
- Render breadcrumbs in PageHeader component
- Fix regex to prevent trailing hyphens/underscores in validation

+++ backend/app/schemas/budget.py
@@ -135,6 +135,6 @@ class BudgetSummaryResponse(CamelCaseModel):
     total_actual: Decimal
     total_variance: Decimal
     total_change_orders: int
-    approved_change_orders: Decimal
+    approved_change_orders: int
     line_item_count: int
     cost_entry_count: int
